Refactor tracking-preview and tracking-page so they don't repeat themselves and re-use new components
